% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaAttrs.R
\name{addEdgeColors}
\alias{addEdgeColors}
\title{Add edge colors by an attribute or by specifying a single color.}
\usage{
addEdgeColors(sigmaObj, oneColor = NULL, colorAttr = NULL,
  colorPal = "Set2")
}
\arguments{
\item{sigmaObj}{A Sigma object - created using the sigmaFromIgraph function}

\item{oneColor}{A single color to color all of the nodes (hex format)}

\item{colorAttr}{An attribute from the original igraph nodes to color the nodes by}

\item{colorPal}{The color palatte to use - only used if colorAttr is specified}
}
\description{
Add edge colors by an attribute or by specifying a single color.
}
