% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaAttrs.R
\name{addEdgeSize}
\alias{addEdgeSize}
\title{Add edge size by specifying an attribute (usually weight) or a single edge size for all edges.}
\usage{
addEdgeSize(sigmaObj, sizeAttr = NULL, minSize = 1, maxSize = 5,
  oneSize = NULL)
}
\arguments{
\item{sigmaObj}{A Sigma object - created using the sigmaFromIgraph function}

\item{sizeAttr}{The attribute to use to create edge size (width)}

\item{minSize}{The minimum size of the edges (for scaling)}

\item{maxSize}{The maximum size of the edges (for scaling)}

\item{oneSize}{A single size to use for all edges}
}
\description{
Add edge size by specifying an attribute (usually weight) or a single edge size for all edges.
}
