% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaAttrs.R
\name{addInteraction}
\alias{addInteraction}
\title{Add/modify interactivity of the visualization}
\usage{
addInteraction(sigmaObj, neighborEvent = "onClick")
}
\arguments{
\item{sigmaObj}{A Sigma object - created using the sigmaFromIgraph function}

\item{neighborEvent}{Enable/disable event that highlights a node's neighbors.  Can either be onClick, onHover, or None.}
}
\description{
Add/modify interactivity of the visualization
}
