% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaAttrs.R
\name{addNodeColors}
\alias{addNodeColors}
\title{Add node colors by an attribute or by specifying a single color.}
\usage{
addNodeColors(sigmaObj, oneColor = NULL, colorAttr = NULL,
  colorPal = "Dark2")
}
\arguments{
\item{sigmaObj}{A Sigma object - created using the sigmaFromIgraph function}

\item{oneColor}{A single color to color all of the nodes (hex format)}

\item{colorAttr}{An attribute from the original igraph nodes to color the nodes by}

\item{colorPal}{The color palatte to use - only used if colorAttr is specified}
}
\description{
Add node colors by an attribute or by specifying a single color.
}
