% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaAttrs.R
\name{addNodeLabels}
\alias{addNodeLabels}
\title{Add node labels by specifying an attribute from the igraph object.}
\usage{
addNodeLabels(sigmaObj, labelAttr = NULL)
}
\arguments{
\item{sigmaObj}{A Sigma object - created using the sigmaFromIgraph function}

\item{labelAttr}{The attribute to use to create node labels}
}
\description{
Add node labels by specifying an attribute from the igraph object.
}
