% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaShiny.R
\name{renderSigmaNet}
\alias{renderSigmaNet}
\title{Render a 'sigmaNet' visualization in 'Shiny'}
\usage{
renderSigmaNet(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that creates a 'sigmaNet' visualization}

\item{env}{Defaults to parent.frame() - see 'Shiny' docs for more info}

\item{quoted}{Defaults to FALSE - see 'Shiny docs for more info}
}
\description{
Render a 'sigmaNet' visualization in 'Shiny'
}
