% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaAttrs.R
\name{saveSigma}
\alias{saveSigma}
\title{Save sigma widget as html - a wrapper for saveWidget()}
\usage{
saveSigma(sigmaObj, fileName = NULL)
}
\arguments{
\item{sigmaObj}{A Sigma object - created using the sigmaFromIgraph function}

\item{fileName}{A name for your html output (with or without .html at the end)}
}
\description{
Save sigma widget as html - a wrapper for saveWidget()
}
