% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaNet.R
\name{sigmaNet}
\alias{sigmaNet}
\title{Draw a network in sigmajs}
\usage{
sigmaNet(graph, layout = NULL, nodeSizeMetric = "degree",
  nodeLabels = NULL, nodeColorAttr = NULL, edgeColorAttr = NULL,
  nodeColorPalette = "Dark2", edgeColorPalette = "Set2",
  edgeWeightAttr = NULL, minNodeSize = 1, maxNodeSize = 8,
  minEdgeSize = 1, maxEdgeSize = 1, nodeColor = "#3182bd",
  edgeColor = "#636363", saveAs = NULL, showInViewer = TRUE,
  width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{graph}{An igraph object}

\item{layout}{The output of one of the igraph layout functions.  If not provided, layout_nicely() will be used (note, this will slow things down).}

\item{nodeSizeMetric}{The metric to use when sizing the nodes.  Options are: degree, closeness, betweenness, pageRank, or eigenCentrality.}

\item{nodeColorPalette}{??}

\item{edgeWeightAttr}{Which attribute to use as edge weights.  Defaults to NULL to render all edges the same size.}

\item{minNodeSize}{The minimum size of graph nodes (defaults to 1)}

\item{maxNodeSize}{The maximum size of graph nodes (defaults to 8)}

\item{minEdgeSize}{The minimum size of graph edges (defaults to 1)}

\item{maxEdgeSize}{The maxiumum size of graph edges (defaults to 1)}

\item{nodeColor}{A color name or hex string specifying the color for all nodes}

\item{edgeColor}{A color name of hex string specifying the color for all edges}

\item{nodeLabel}{Which attribute to use as the node labels.  Default will be the "names" assigned by igraph - you probably don't need to change this.}
}
\description{
Draw a network in sigmajs
}
