% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaShiny.R
\name{sigmaNetOutput}
\alias{sigmaNetOutput}
\title{Create a UI element for a 'sigmaNet' visualization in 'Shiny'}
\usage{
sigmaNetOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{The ID of the UI element}

\item{width}{The width of the UI element}

\item{height}{The height of the UI element}
}
\description{
Create a UI element for a 'sigmaNet' visualization in 'Shiny'
}
