% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaAttrs.R
\name{addListener}
\alias{addListener}
\title{Add a "listener" to report data from a 'sigmaNet' object in 'Shiny' back to the R session.}
\usage{
addListener(sigmaObj, listener)
}
\arguments{
\item{sigmaObj}{A 'sigmaNet' object - created using the 'sigmaFromIgraph' function}

\item{listener}{Either "clickNode" to listen to node clicks or "hoverNode" to listen to node hover}
}
\description{
Add a "listener" to report data from a 'sigmaNet' object in 'Shiny' back to the R session.
}
