% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaAttrs.R
\name{addNodeSize}
\alias{addNodeSize}
\title{Add node size by either specifying a metric, or supplying your own vector of sizes.}
\usage{
addNodeSize(sigmaObj, minSize = 1, maxSize = 3, sizeMetric = "degree",
  sizeVector = NULL, oneSize = NULL)
}
\arguments{
\item{sigmaObj}{A Sigma object - created using the sigmaFromIgraph function}

\item{minSize}{The minimum node size on the graph (for scaling)}

\item{maxSize}{The maximum node size on the graph (for scaling)}

\item{sizeVector}{An optional vector with the sizes for each node}

\item{oneSize}{A single size to use for all nodes}

\item{sizeAttr}{An attribute from the original igraph nodes to size the nodes by}
}
\description{
Add node size by either specifying a metric, or supplying your own vector of sizes.
}
