% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaNet.R
\name{sigmaNet}
\alias{sigmaNet}
\title{Draw a network in sigmajs}
\usage{
sigmaNet(graph, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{graph}{An igraph object}
}
\description{
Draw a network in sigmajs
}
